Add new Student instances and update Course
The main function in Main.java is now creating more instances of the Student class. Besides updating the Student class with the Builder pattern, it modifies the Course instance to include these new students. This changes the persons available in the Java Bootcamp course, which now includes Aljoscha, Jonas, Gökhan, and Daniel.